requirements.txt
replicate>=0.24.0
python-dotenv>=1.0.0

